chore: redistribute background context files
