+++ packages/apps/human-app/frontend/src/shared/contexts/background-context.tsx
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+
+export interface BackgroundContextProps {
+  backgroundColor: string;
+  setWhiteBackground: () => void;
+  setGrayBackground: () => void;
+}
+
+export const BackgroundContext = createContext<
+  BackgroundContextProps | undefined
+>(undefined);
